Revert "Merge pull request #144 from ejc3/fcvm-podman-cache"
This reverts commit f4dfb1ce2308218e02e4051346341615a204906d, reversing
changes made to 6c6e470eca521d0bf168bbf57ba787ccde8c53c7.
