+++ fc-agent/src/main.rs
@@ -1135,178 +1135,6 @@ fn notify_container_exit(exit_code: i32) {
     }
 }
 
-/// Get the digest of a pulled image using podman inspect.
-/// Returns the digest string (e.g., "sha256:abc123...")
-async fn get_image_digest(image: &str) -> Result<String> {
-    let output = Command::new("podman")
-        .args(["image", "inspect", "--format", "{{.Digest}}", image])
-        .output()
-        .await
-        .context("running podman image inspect")?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        anyhow::bail!("podman image inspect failed: {}", stderr);
-    }
-
-    let digest = String::from_utf8_lossy(&output.stdout).trim().to_string();
-    Ok(digest)
-}
-
-/// Notify host that image is loaded and wait for cache creation acknowledgment.
-///
-/// Sends "cache-ready:{digest}\n" to host and waits for "cache-ack\n" response.
-/// The host will pause the VM, create a snapshot, resume, then send ack.
-/// This blocks indefinitely until ack is received (no timeout).
-fn notify_cache_ready_and_wait(digest: &str) -> bool {
-    // Create vsock socket
-    let fd = unsafe { libc::socket(libc::AF_VSOCK, libc::SOCK_STREAM, 0) };
-    if fd < 0 {
-        eprintln!(
-            "[fc-agent] WARNING: failed to create vsock socket for cache: {}",
-            std::io::Error::last_os_error()
-        );
-        return false;
-    }
-
-    // Build sockaddr_vm structure
-    let addr = libc::sockaddr_vm {
-        svm_family: libc::AF_VSOCK as u16,
-        svm_reserved1: 0,
-        svm_port: STATUS_VSOCK_PORT,
-        svm_cid: HOST_CID,
-        svm_zero: [0u8; 4],
-    };
-
-    // Connect to host
-    let result = unsafe {
-        libc::connect(
-            fd,
-            &addr as *const libc::sockaddr_vm as *const libc::sockaddr,
-            std::mem::size_of::<libc::sockaddr_vm>() as u32,
-        )
-    };
-
-    if result < 0 {
-        let err = std::io::Error::last_os_error();
-        unsafe { libc::close(fd) };
-        eprintln!(
-            "[fc-agent] WARNING: failed to connect vsock for cache: {}",
-            err
-        );
-        return false;
-    }
-
-    // Send cache-ready message
-    let msg = format!("cache-ready:{}\n", digest);
-    let written = unsafe { libc::write(fd, msg.as_ptr() as *const libc::c_void, msg.len()) };
-
-    if written != msg.len() as isize {
-        eprintln!("[fc-agent] WARNING: failed to send cache-ready message");
-        unsafe { libc::close(fd) };
-        return false;
-    }
-
-    eprintln!("[fc-agent] sent cache-ready:{}, waiting for ack...", digest);
-
-    // Wait for cache-ack response with timeout (handles snapshot restore case)
-    // After restore, the vsock connection is dead and poll will timeout
-    let mut buf = [0u8; 64];
-    let mut total_read = 0;
-    const POLL_TIMEOUT_MS: i32 = 100; // 100ms timeout
-
-    // Set socket to non-blocking to prevent read() from blocking after restore
-    // After snapshot restore, the kernel might think data is available but read() would block
-    unsafe {
-        let flags = libc::fcntl(fd, libc::F_GETFL);
-        if flags >= 0 {
-            libc::fcntl(fd, libc::F_SETFL, flags | libc::O_NONBLOCK);
-        }
-    }
-
-    loop {
-        // Use poll() to check if data is available with timeout
-        let mut pfd = libc::pollfd {
-            fd,
-            events: libc::POLLIN,
-            revents: 0,
-        };
-
-        let poll_result = unsafe { libc::poll(&mut pfd, 1, POLL_TIMEOUT_MS) };
-
-        if poll_result < 0 {
-            // poll error
-            eprintln!(
-                "[fc-agent] cache-ack poll error: {}",
-                std::io::Error::last_os_error()
-            );
-            unsafe { libc::close(fd) };
-            return false;
-        }
-
-        if poll_result == 0 {
-            // Timeout - likely restored from snapshot with dead connection
-            eprintln!("[fc-agent] cache-ack poll timeout (restored from snapshot?)");
-            unsafe { libc::close(fd) };
-            return false;
-        }
-
-        // Check for hangup/error
-        if pfd.revents & (libc::POLLHUP | libc::POLLERR) != 0 {
-            eprintln!("[fc-agent] cache-ack connection closed or error");
-            unsafe { libc::close(fd) };
-            return false;
-        }
-
-        // Data available (POLLIN set), do non-blocking read
-        let n = unsafe {
-            libc::read(
-                fd,
-                buf[total_read..].as_mut_ptr() as *mut libc::c_void,
-                buf.len() - total_read,
-            )
-        };
-
-        if n < 0 {
-            let err = std::io::Error::last_os_error();
-            if err.kind() == std::io::ErrorKind::WouldBlock {
-                // Non-blocking read returned EAGAIN/EWOULDBLOCK - no data actually available
-                // This can happen after snapshot restore when poll returns but data isn't really there
-                eprintln!("[fc-agent] cache-ack read would block (likely restored from snapshot)");
-                unsafe { libc::close(fd) };
-                return false;
-            }
-            eprintln!("[fc-agent] cache-ack read error: {}", err);
-            unsafe { libc::close(fd) };
-            return false;
-        }
-
-        if n == 0 {
-            // Connection closed
-            eprintln!("[fc-agent] cache-ack connection closed");
-            unsafe { libc::close(fd) };
-            return false;
-        }
-
-        total_read += n as usize;
-
-        // Check if we received "cache-ack\n"
-        let received = std::str::from_utf8(&buf[..total_read]).unwrap_or("");
-        if received.contains("cache-ack") {
-            eprintln!("[fc-agent] ✓ received cache-ack from host");
-            unsafe { libc::close(fd) };
-            return true;
-        }
-
-        // Prevent buffer overflow
-        if total_read >= buf.len() {
-            eprintln!("[fc-agent] cache-ack buffer overflow, giving up");
-            unsafe { libc::close(fd) };
-            return false;
-        }
-    }
-}
-
 /// Shutdown the VM with the given exit code.
 ///
 /// This function handles the shutdown sequence:
@@ -2015,64 +1843,14 @@ async fn run_agent() -> Result<()> {
         plan.image.clone()
     };
 
-    // Notify host that image is ready for caching
-    // For registry images, get the digest from podman
-    // For OCI archives, use the image name as identifier
-    if !image_ref.starts_with("oci-archive:") {
-        // Registry image - get digest from podman
-        match get_image_digest(&plan.image).await {
-            Ok(digest) => {
-                eprintln!("[fc-agent] image digest: {}", digest);
-                if notify_cache_ready_and_wait(&digest) {
-                    eprintln!("[fc-agent] ✓ cache ready notification acknowledged");
-                } else {
-                    eprintln!(
-                        "[fc-agent] WARNING: cache-ready handshake failed, continuing anyway"
-                    );
-                }
-            }
-            Err(e) => {
-                eprintln!("[fc-agent] WARNING: failed to get image digest: {:?}", e);
-                eprintln!("[fc-agent] continuing without cache notification");
-            }
-        }
-    } else {
-        // OCI archive - notify with image name as identifier
-        eprintln!("[fc-agent] using OCI archive, notifying cache-ready with image name");
-        if notify_cache_ready_and_wait(&plan.image) {
-            eprintln!("[fc-agent] ✓ cache ready notification acknowledged");
-        } else {
-            eprintln!("[fc-agent] WARNING: cache-ready handshake failed, continuing anyway");
-        }
-    }
-
     eprintln!("[fc-agent] launching container: {}", image_ref);
 
     // Build Podman args (used for both TTY and non-TTY modes)
-    //
-    // CRITICAL: --cgroups=split is REQUIRED for snapshot restore.
-    //
-    // Root cause: After Firecracker snapshot restore, systemd's timer/event loop
-    // gets confused because the VM clock jumps from snapshot-time to current-time.
-    // This is a known class of issues (see systemd/systemd#23032) where:
-    //   1. systemd's CLOCK_REALTIME-based timers break on clock discontinuity
-    //   2. epoll_wait() and timerfd handling in the event loop become confused
-    //   3. The watchdog/keepalive pings don't get sent correctly
-    //   4. D-Bus calls to systemd time out waiting for responses
-    //
-    // When crun tries to create cgroups via sd-bus, systemd never responds:
-    //   "OCI runtime error: crun: sd-bus call: Connection timed out"
-    //
-    // Solution: --cgroups=split uses cgroups v2 directly via cgroupfs, bypassing
-    // the broken D-Bus/systemd communication. This gives us:
-    // - Full cgroup v2 functionality (resource limits, isolation)
-    // - No dependency on systemd's event loop
     let mut podman_args = vec![
         "podman".to_string(),
         "run".to_string(),
         "--rm".to_string(),
         "--network=host".to_string(),
-        "--cgroups=split".to_string(),
         "--ulimit".to_string(),
         "nofile=65536:65536".to_string(),
     ];
