+++ src/commands/common.rs
@@ -9,14 +9,10 @@ use anyhow::{Context, Result};
 use tokio::task::JoinHandle;
 use tracing::{debug, info, warn};
 
-use std::path::PathBuf;
-
 use crate::{
     firecracker::VmManager,
-    network::{BridgedNetwork, NetworkConfig, NetworkManager, SlirpNetwork},
-    paths,
+    network::{NetworkConfig, NetworkManager},
     state::{StateManager, VmState, VmStatus},
-    storage::DiskManager,
 };
 
 /// Vsock base port for volume servers (used by both podman and snapshot commands)
@@ -198,425 +194,4 @@ pub async fn cleanup_vm(
     } else {
         info!(vm_id = %vm_id, "cleaned up VM data directory");
     }
-}
-
-/// Memory backend configuration for snapshot restore
-pub enum MemoryBackend {
-    /// Load memory directly from file (used by podman cache restore)
-    File { memory_path: PathBuf },
-    /// Use UFFD server for on-demand page loading (used by snapshot clones)
-    Uffd { socket_path: PathBuf },
-}
-
-/// Configuration for restoring a VM from a snapshot
-pub struct SnapshotRestoreConfig {
-    /// VM state path (vmstate.bin)
-    pub vmstate_path: PathBuf,
-    /// Memory backend configuration
-    pub memory_backend: MemoryBackend,
-    /// Source disk for CoW copy
-    pub source_disk_path: PathBuf,
-    /// Original VM ID for vsock socket path redirect (from original cache creation)
-    pub original_vm_id: String,
-    /// Snapshot VM ID for disk path redirect (the VM that was snapshotted)
-    /// This is needed because disk paths are patched during cache restore,
-    /// so vmstate.bin has a different VM ID for disk than for vsock.
-    pub snapshot_vm_id: Option<String>,
-}
-
-/// Restore a VM from a snapshot
-///
-/// This is the core snapshot restore logic shared by:
-/// - `fcvm snapshot run` (clone with UFFD memory sharing)
-/// - `fcvm podman run` with cache hit (direct file load)
-///
-/// Both paths use identical Firecracker setup, the only differences are:
-/// - Memory backend: UFFD vs File
-/// - Snapshot source: snapshots/{name} vs podman-cache/{hash}
-#[allow(clippy::too_many_arguments)]
-pub async fn restore_from_snapshot(
-    vm_id: &str,
-    vm_name: &str,
-    data_dir: &Path,
-    socket_path: &Path,
-    restore_config: &SnapshotRestoreConfig,
-    network_config: &NetworkConfig,
-    network: &mut dyn NetworkManager,
-    state_manager: &StateManager,
-    vm_state: &mut VmState,
-) -> Result<(VmManager, Option<tokio::process::Child>)> {
-    let vm_dir = data_dir.join("disks");
-
-    // Configure namespace isolation if network provides one
-    let mut holder_child: Option<tokio::process::Child> = None;
-    let mut holder_pid_for_post_start: Option<u32> = None;
-    let mut vm_manager = VmManager::new(vm_id.to_string(), socket_path.to_path_buf(), None);
-    vm_manager.set_vm_name(vm_name.to_string());
-
-    // rootfs_path is set by either the bridged or rootless branch
-    let rootfs_path: PathBuf;
-
-    if let Some(bridged_net) = network.as_any().downcast_ref::<BridgedNetwork>() {
-        if let Some(ns_id) = bridged_net.namespace_id() {
-            info!(namespace = %ns_id, "configuring VM to run in network namespace");
-            vm_manager.set_namespace(ns_id.to_string());
-        }
-
-        // For bridged mode, create disk
-        let disk_manager = DiskManager::new(
-            vm_id.to_string(),
-            restore_config.source_disk_path.clone(),
-            vm_dir.clone(),
-        );
-
-        rootfs_path = disk_manager
-            .create_cow_disk()
-            .await
-            .context("creating CoW disk from snapshot")?;
-
-        info!(
-            rootfs = %rootfs_path.display(),
-            source_disk = %restore_config.source_disk_path.display(),
-            "CoW disk prepared from snapshot"
-        );
-    } else if let Some(slirp_net) = network.as_any().downcast_ref::<SlirpNetwork>() {
-        // Rootless mode: spawn holder process and set up namespace via nsenter
-        // OPTIMIZATION: Parallelize disk creation with network setup
-
-        // Step 1: Spawn holder process (keeps namespace alive)
-        let holder_cmd = slirp_net.build_holder_command();
-        info!(cmd = ?holder_cmd, "spawning namespace holder for rootless networking");
-
-        let retry_deadline = std::time::Instant::now() + std::time::Duration::from_secs(5);
-        let mut attempt = 0u32;
-
-        let (mut child, holder_pid) = loop {
-            attempt += 1;
-
-            let mut child = tokio::process::Command::new(&holder_cmd[0])
-                .args(&holder_cmd[1..])
-                .stdout(std::process::Stdio::null())
-                .stderr(std::process::Stdio::null())
-                .spawn()
-                .context("spawning namespace holder process")?;
-
-            let holder_pid = child.id().context("getting holder process PID")?;
-            if attempt > 1 {
-                info!(
-                    holder_pid = holder_pid,
-                    attempt = attempt,
-                    "namespace holder started (retry)"
-                );
-            } else {
-                info!(holder_pid = holder_pid, "namespace holder started");
-            }
-
-            // Wait for namespace to be ready by checking uid_map
-            let namespace_ready = crate::utils::wait_for_namespace_ready(
-                holder_pid,
-                std::time::Duration::from_millis(500),
-            )
-            .await;
-
-            if namespace_ready {
-                break (child, holder_pid);
-            }
-
-            // Namespace not ready, kill holder and retry
-            let _ = child.kill().await;
-            if std::time::Instant::now() < retry_deadline {
-                warn!(
-                    holder_pid = holder_pid,
-                    attempt = attempt,
-                    "namespace not ready, retrying holder creation..."
-                );
-                tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-            } else {
-                anyhow::bail!(
-                    "namespace not ready after {} attempts (holder PID {})",
-                    attempt,
-                    holder_pid
-                );
-            }
-        };
-
-        // Step 2: Run disk creation and network setup IN PARALLEL
-        let setup_script = slirp_net.build_setup_script();
-        let nsenter_prefix = slirp_net.build_nsenter_prefix(holder_pid);
-        let tap_device = network_config.tap_device.clone();
-
-        // Disk creation task
-        let source_disk = restore_config.source_disk_path.clone();
-        let disk_task = async {
-            let disk_manager =
-                DiskManager::new(vm_id.to_string(), source_disk.clone(), vm_dir.clone());
-
-            let rootfs_path = disk_manager
-                .create_cow_disk()
-                .await
-                .context("creating CoW disk from snapshot")?;
-
-            info!(
-                rootfs = %rootfs_path.display(),
-                source_disk = %source_disk.display(),
-                "CoW disk prepared from snapshot"
-            );
-
-            Ok::<_, anyhow::Error>(rootfs_path)
-        };
-
-        // Network setup task
-        let network_task = async {
-            const MAX_NS_WAIT: std::time::Duration = std::time::Duration::from_millis(1000);
-            const NS_POLL_INTERVAL: std::time::Duration = std::time::Duration::from_millis(5);
-            let ns_poll_start = std::time::Instant::now();
-
-            info!(holder_pid = holder_pid, "running network setup via nsenter");
-            loop {
-                // Verify holder is still alive before attempting nsenter
-                if !crate::utils::is_process_alive(holder_pid) {
-                    anyhow::bail!(
-                        "holder process (PID {}) died before network setup could run",
-                        holder_pid
-                    );
-                }
-
-                let output = tokio::process::Command::new(&nsenter_prefix[0])
-                    .args(&nsenter_prefix[1..])
-                    .arg("bash")
-                    .arg("-c")
-                    .arg(&setup_script)
-                    .output()
-                    .await
-                    .context("running network setup via nsenter")?;
-
-                if output.status.success() {
-                    debug!("namespace ready after {:?}", ns_poll_start.elapsed());
-                    break;
-                }
-
-                // Check if it's a namespace-not-ready error (retry) vs permanent error (fail)
-                let stderr = String::from_utf8_lossy(&output.stderr);
-                if stderr.contains("Invalid argument") || stderr.contains("No such process") {
-                    if ns_poll_start.elapsed() > MAX_NS_WAIT {
-                        anyhow::bail!(
-                            "namespace not ready after {:?}: {}",
-                            ns_poll_start.elapsed(),
-                            stderr
-                        );
-                    }
-                    tokio::time::sleep(NS_POLL_INTERVAL).await;
-                    continue;
-                }
-
-                // Permanent error
-                anyhow::bail!("network setup failed: {}", stderr);
-            }
-
-            // Verify TAP device was created successfully
-            let verify_cmd = format!("ip link show {} >/dev/null 2>&1", tap_device);
-            let verify_output = tokio::process::Command::new(&nsenter_prefix[0])
-                .args(&nsenter_prefix[1..])
-                .arg("bash")
-                .arg("-c")
-                .arg(&verify_cmd)
-                .status()
-                .await
-                .context("verifying TAP device")?;
-
-            if !verify_output.success() {
-                anyhow::bail!(
-                    "TAP device '{}' not found after network setup - setup may have failed silently",
-                    tap_device
-                );
-            }
-            debug!(tap_device = %tap_device, "TAP device verified");
-
-            Ok::<_, anyhow::Error>(())
-        };
-
-        // Run both tasks in parallel
-        let (disk_result, network_result) = tokio::join!(disk_task, network_task);
-
-        // Handle errors - kill holder child if either fails
-        if let Err(e) = &disk_result {
-            let _ = child.kill().await;
-            return Err(anyhow::anyhow!("disk creation failed: {}", e));
-        }
-        if let Err(e) = &network_result {
-            let _ = child.kill().await;
-            return Err(anyhow::anyhow!("network setup failed: {}", e));
-        }
-
-        rootfs_path = disk_result?;
-        network_result?;
-
-        info!(
-            holder_pid = holder_pid,
-            "parallel disk + network setup complete"
-        );
-
-        // Step 3: Set namespace paths for pre_exec setns
-        vm_manager.set_user_namespace_path(PathBuf::from(format!("/proc/{}/ns/user", holder_pid)));
-        vm_manager.set_net_namespace_path(PathBuf::from(format!("/proc/{}/ns/net", holder_pid)));
-
-        // Store holder_pid in state for health checks
-        vm_state.holder_pid = Some(holder_pid);
-        holder_pid_for_post_start = Some(holder_pid);
-
-        holder_child = Some(child);
-    } else {
-        // Unknown network type - should not happen
-        anyhow::bail!("Unknown network type - must be either BridgedNetwork or SlirpNetwork");
-    }
-
-    // Configure mount namespace isolation for path redirects
-    // We need to redirect BOTH:
-    // 1. original_vm_id - for vsock paths in vmstate.bin (original cache VM)
-    // 2. snapshot_vm_id - for disk paths in vmstate.bin (snapshotted VM, if different)
-    let mut baseline_dirs = vec![paths::vm_runtime_dir(&restore_config.original_vm_id)];
-    if let Some(ref snapshot_vm_id) = restore_config.snapshot_vm_id {
-        if snapshot_vm_id != &restore_config.original_vm_id {
-            baseline_dirs.push(paths::vm_runtime_dir(snapshot_vm_id));
-        }
-    }
-    info!(
-        baseline_dirs = ?baseline_dirs.iter().map(|p| p.display().to_string()).collect::<Vec<_>>(),
-        clone_dir = %data_dir.display(),
-        "enabling mount namespace for path isolation"
-    );
-    vm_manager.set_mount_redirects(baseline_dirs, data_dir.to_path_buf());
-
-    let firecracker_bin = find_firecracker()?;
-
-    vm_manager
-        .start(&firecracker_bin, None)
-        .await
-        .context("starting Firecracker")?;
-
-    // For rootless mode with slirp4netns: post_start starts slirp4netns in the namespace
-    let vm_pid = vm_manager.pid()?;
-    let post_start_pid = holder_pid_for_post_start.unwrap_or(vm_pid);
-    network
-        .post_start(post_start_pid)
-        .await
-        .context("post-start network setup")?;
-
-    let client = vm_manager.client()?;
-
-    // Load snapshot with configured memory backend and network override
-    use crate::firecracker::api::{
-        DrivePatch, MemBackend, NetworkOverride, SnapshotLoad, VmState as ApiVmState,
-    };
-
-    let mem_backend = match &restore_config.memory_backend {
-        MemoryBackend::File { memory_path } => {
-            info!(
-                memory = %memory_path.display(),
-                "loading snapshot with File backend"
-            );
-            MemBackend {
-                backend_type: "File".to_string(),
-                backend_path: memory_path.display().to_string(),
-            }
-        }
-        MemoryBackend::Uffd { socket_path } => {
-            info!(
-                uffd_socket = %socket_path.display(),
-                "loading snapshot with UFFD backend"
-            );
-            MemBackend {
-                backend_type: "Uffd".to_string(),
-                backend_path: socket_path.display().to_string(),
-            }
-        }
-    };
-
-    // Timing instrumentation: measure snapshot load operation
-    let load_start = std::time::Instant::now();
-    client
-        .load_snapshot(SnapshotLoad {
-            snapshot_path: restore_config.vmstate_path.display().to_string(),
-            mem_backend,
-            enable_diff_snapshots: Some(false),
-            resume_vm: Some(false), // Update devices before resume
-            network_overrides: Some(vec![NetworkOverride {
-                iface_id: "eth0".to_string(),
-                host_dev_name: network_config.tap_device.clone(),
-            }]),
-        })
-        .await
-        .context("loading snapshot")?;
-    let load_duration = load_start.elapsed();
-    info!(
-        duration_ms = load_duration.as_millis(),
-        "snapshot load completed"
-    );
-
-    // Timing instrumentation: measure disk patch operation
-    let patch_start = std::time::Instant::now();
-    client
-        .patch_drive(
-            "rootfs",
-            DrivePatch {
-                drive_id: "rootfs".to_string(),
-                path_on_host: Some(rootfs_path.display().to_string()),
-                rate_limiter: None,
-            },
-        )
-        .await
-        .context("retargeting rootfs drive")?;
-    let patch_duration = patch_start.elapsed();
-    info!(
-        duration_ms = patch_duration.as_millis(),
-        "disk patch completed"
-    );
-
-    // Signal fc-agent to flush ARP cache via MMDS restore-epoch update
-    let restore_epoch = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .context("system time before Unix epoch")?
-        .as_secs();
-
-    client
-        .put_mmds(serde_json::json!({
-            "latest": {
-                "host-time": chrono::Utc::now().timestamp().to_string(),
-                "restore-epoch": restore_epoch.to_string()
-            }
-        }))
-        .await
-        .context("updating MMDS with restore-epoch")?;
-    info!(
-        restore_epoch = restore_epoch,
-        "signaled fc-agent to flush ARP via MMDS"
-    );
-
-    // Timing instrumentation: measure VM resume operation
-    let resume_start = std::time::Instant::now();
-    client
-        .patch_vm_state(ApiVmState {
-            state: "Resumed".to_string(),
-        })
-        .await
-        .context("resuming VM after snapshot load")?;
-    let resume_duration = resume_start.elapsed();
-    info!(
-        duration_ms = resume_duration.as_millis(),
-        total_snapshot_ms = (load_duration + patch_duration + resume_duration).as_millis(),
-        "VM resume completed"
-    );
-
-    // Store fcvm process PID (not Firecracker PID)
-    vm_state.pid = Some(std::process::id());
-
-    // Track original vsock vm_id for future snapshots
-    // When this VM is later snapshotted, clones need to use this original_vm_id
-    // for vsock redirect because vmstate.bin stores paths from this vm
-    vm_state.config.original_vsock_vm_id = Some(restore_config.original_vm_id.clone());
-
-    // Save VM state with complete network configuration
-    save_vm_state_with_network(state_manager, vm_state, network_config).await?;
-
-    Ok((vm_manager, holder_child))
 }
