+++ src/firecracker/mod.rs
@@ -1,7 +1,5 @@
 pub mod api;
-pub mod config;
 pub mod vm;
 
 pub use api::FirecrackerClient;
-pub use config::{FirecrackerConfig, NetworkMode as FcNetworkMode};
 pub use vm::VmManager;
